feat(Button): create a product CTA button.
- add the product CTA button and import in the Zx7 section
- improve on the white space in the section.
